fix(clusters): switch from drain to ineligible system (#1102)
<!-- Please make sure there is an issue that this PR is correlated to. -->
Fixes RVTEE-574

**Manually tested**
## Changes

<!-- If there are frontend changes, please include screenshots. -->

+++ lib/chirp-workflow/core/src/worker.rs
@@ -127,7 +127,7 @@ impl Worker {
 				async move {
 					// Sleep until deadline
 					if let Some(wake_deadline_ts) = wake_deadline_ts {
-						util::time::sleep_until_ts(wake_deadline_ts).await;
+						util::time::sleep_until_ts(wake_deadline_ts as u64).await;
 					}
 
 					ctx.run().await;
